Added small changes for Katacoda
Added small changes in order to get it better suitable for Katacoda.
