+++ java/src/main/java/lesson01/exercise/Hello.java
@@ -1,0 +1,4 @@
+package lesson01.exercise;
+
+public class Hello {
+}
