+++ java/src/main/java/lesson02/exercise/Hello.java
@@ -1,0 +1,4 @@
+package lesson02.exercise;
+
+public class Hello {
+}
