+++ java/src/main/java/lesson03/exercise/RequestBuilderCarrier.java
@@ -1,0 +1,4 @@
+package lesson03.exercise;
+
+public class RequestBuilderCarrier {
+}
