+++ java/src/main/java/lesson03/solution/RequestBuilderCarrier.java
@@ -1,0 +1,24 @@
+package lesson03.solution;
+
+import java.util.Iterator;
+import java.util.Map;
+
+import okhttp3.Request;
+
+public class RequestBuilderCarrier implements io.opentracing.propagation.TextMap {
+    private final Request.Builder builder;
+
+    RequestBuilderCarrier(Request.Builder builder) {
+        this.builder = builder;
+    }
+
+    @Override
+    public Iterator<Map.Entry<String, String>> iterator() {
+        throw new UnsupportedOperationException("carrier is write-only");
+    }
+
+    @Override
+    public void put(String key, String value) {
+        builder.addHeader(key, value);
+    }
+}
