+++ java/src/main/java/lesson04/exercise/Formatter.java
@@ -1,0 +1,4 @@
+package lesson04.exercise;
+
+public class Formatter {
+}
