+++ java/src/main/java/lesson04/solution/Formatter.java
@@ -1,13 +1,5 @@
 package lesson04.solution;
 
-import com.google.common.collect.ImmutableMap;
-
-import io.dropwizard.Application;
-import io.dropwizard.Configuration;
-import io.dropwizard.setup.Environment;
-import io.opentracing.Scope;
-import io.opentracing.Tracer;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
@@ -16,6 +8,13 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 
+import com.google.common.collect.ImmutableMap;
+
+import io.dropwizard.Application;
+import io.dropwizard.Configuration;
+import io.dropwizard.setup.Environment;
+import io.opentracing.Scope;
+import io.opentracing.Tracer;
 import lib.Tracing;
 
 public class Formatter extends Application<Configuration> {
