+++ java/src/main/java/lesson04/solution/Hello.java
@@ -1,19 +1,18 @@
 package lesson04.solution;
 
-import io.opentracing.Scope;
-import io.opentracing.propagation.Format.Builtin;
-import io.opentracing.tag.Tags;
-
 import java.io.IOException;
-import java.util.Iterator;
-import java.util.Map.Entry;
+
 import com.google.common.collect.ImmutableMap;
 import com.uber.jaeger.Tracer;
+
+import io.opentracing.Scope;
+import io.opentracing.propagation.Format;
+import io.opentracing.tag.Tags;
+import lib.Tracing;
+import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
-import okhttp3.HttpUrl;
-import lib.Tracing;
 
 public class Hello {
 
@@ -25,25 +24,6 @@ public class Hello {
         this.client = new OkHttpClient();
     }
 
-    private static class RequestBuilderCarrier implements io.opentracing.propagation.TextMap {
-        private final Request.Builder builder;
-
-        RequestBuilderCarrier(Request.Builder builder) {
-            this.builder = builder;
-        }
-
-		@Override
-		public Iterator<Entry<String, String>> iterator() {
-			throw new UnsupportedOperationException("carrier is write-only");
-		}
-
-		@Override
-		public void put(String key, String value) {
-			builder.addHeader(key, value);
-		}
-
-    }
-
     private String getHttp(int port, String path, String param, String value) {
         try {
             HttpUrl url = new HttpUrl.Builder().scheme("http").host("localhost").port(port).addPathSegment(path)
@@ -53,7 +33,7 @@ public class Hello {
             Tags.SPAN_KIND.set(tracer.activeSpan(), Tags.SPAN_KIND_CLIENT);
             Tags.HTTP_METHOD.set(tracer.activeSpan(), "GET");
             Tags.HTTP_URL.set(tracer.activeSpan(), url.toString());
-            tracer.inject(tracer.activeSpan().context(), Builtin.HTTP_HEADERS, new RequestBuilderCarrier(requestBuilder));
+            tracer.inject(tracer.activeSpan().context(), Format.Builtin.HTTP_HEADERS, Tracing.requestBuilderCarrier(requestBuilder));
 
             Request request = requestBuilder.build();
             Response response = client.newCall(request).execute();
@@ -100,6 +80,5 @@ public class Hello {
         Tracer tracer = Tracing.init("hello-world");
         new Hello(tracer).sayHello(helloTo, greeting);
         tracer.close();
-        System.exit(0); // okhttpclient sometimes hangs maven otherwise
     }
 }
