+++ java/src/main/java/lib/Tracing.java
@@ -1,31 +1,45 @@
 package lib;
 
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
+
+import javax.ws.rs.core.MultivaluedMap;
+
 import com.uber.jaeger.Configuration;
 import com.uber.jaeger.Configuration.ReporterConfiguration;
 import com.uber.jaeger.Configuration.SamplerConfiguration;
+import com.uber.jaeger.samplers.ConstSampler;
+
 import io.opentracing.Scope;
 import io.opentracing.SpanContext;
 import io.opentracing.Tracer;
 import io.opentracing.propagation.Format;
+import io.opentracing.propagation.TextMap;
 import io.opentracing.propagation.TextMapExtractAdapter;
 import io.opentracing.tag.Tags;
-
-import java.util.HashMap;
-import javax.ws.rs.core.MultivaluedMap;
+import okhttp3.Request;
 
 public final class Tracing {
     private Tracing() {
     }
 
     public static com.uber.jaeger.Tracer init(String service) {
-        SamplerConfiguration samplerConfig = new SamplerConfiguration("const", 1);
-        ReporterConfiguration reporterConfig = new ReporterConfiguration(true, null, null, null, null);
-        Configuration config = new Configuration(service, samplerConfig, reporterConfig);
+        SamplerConfiguration samplerConfig = SamplerConfiguration.fromEnv()
+                .withType(ConstSampler.TYPE)
+                .withParam(1);
+
+        ReporterConfiguration reporterConfig = ReporterConfiguration.fromEnv()
+                .withLogSpans(true);
+
+        Configuration config = new Configuration(service)
+                .withSampler(samplerConfig)
+                .withReporter(reporterConfig);
+
         return (com.uber.jaeger.Tracer) config.getTracer();
     }
 
-    public static Scope startServerSpan(Tracer tracer, javax.ws.rs.core.HttpHeaders httpHeaders,
-            String operationName) {
+    public static Scope startServerSpan(Tracer tracer, javax.ws.rs.core.HttpHeaders httpHeaders, String operationName) {
         // format the headers for extraction
         MultivaluedMap<String, String> rawHeaders = httpHeaders.getRequestHeaders();
         final HashMap<String, String> headers = new HashMap<String, String>();
@@ -46,5 +60,19 @@ public final class Tracing {
         }
         // TODO could add more tags like http.url
         return spanBuilder.withTag(Tags.SPAN_KIND.getKey(), Tags.SPAN_KIND_SERVER).startActive(true);
+    }
+
+    public static TextMap requestBuilderCarrier(final Request.Builder builder) {
+        return new TextMap() {
+            @Override
+            public Iterator<Map.Entry<String, String>> iterator() {
+                throw new UnsupportedOperationException("carrier is write-only");
+            }
+
+            @Override
+            public void put(String key, String value) {
+                builder.addHeader(key, value);
+            }
+        };
     }
 }
